Allow installation on IntelliJ 2024.3.x
Note: pluginSinceBuild is still 241. This changes
the platformVersion back to 241, so that it is built
against this version to guarantee backwards-compatibility.
The incompatibility in PMDProjectComponent is also
fixed (refs #181, #185, #188).

See https://plugins.jetbrains.com/docs/intellij/build-number-ranges.html#multipleIDEVersions

Note: This PR requires most likely #190 so that the plugin works in 2024.3.